+++ packages/headless/src/app/search-thunk-extra-arguments.ts
@@ -1,6 +1,10 @@
 import {SearchAPIClient} from '../api/search/search-api-client';
-import {ThunkExtraArguments} from './thunk-extra-arguments';
+import {ClientThunkExtraArguments} from './thunk-extra-arguments';
 
-export interface SearchThunkExtraArguments extends ThunkExtraArguments {
+export interface SearchThunkExtraArguments
+  extends ClientThunkExtraArguments<SearchAPIClient> {
+  /**
+   * @deprecated - The `searchAPIClient` property is now `apiClient`.
+   */
   searchAPIClient: SearchAPIClient;
 }