+++ packages/headless/src/features/facets/facet-search-set/generic/generic-facet-search-actions.ts
@@ -2,7 +2,6 @@ import {createAction, createAsyncThunk} from '@reduxjs/toolkit';
 import {CategoryFacetSearchRequest} from '../../../../api/search/facet-search/category-facet-search/category-facet-search-request';
 import {FacetSearchResponse} from '../../../../api/search/facet-search/facet-search-response';
 import {SpecificFacetSearchRequest} from '../../../../api/search/facet-search/specific-facet-search/specific-facet-search-request';
-import {AsyncThunkSearchOptions} from '../../../../api/search/search-api-client';
 import {logFacetSearch} from '../../facet-set/facet-set-analytics-actions';
 import {buildSpecificFacetSearchRequest} from '../specific/specific-facet-search-request-builder';
 import {buildCategoryFacetSearchRequest} from '../category/category-facet-search-request-builder';
@@ -16,6 +15,9 @@ import {
   validatePayload,
 } from '../../../../utils/validate-payload';
 import {facetIdDefinition} from '../../generic/facet-actions-validation';
+import {AsyncThunkOptions} from '../../../../app/async-thunk-options';
+import {ClientThunkExtraArguments} from '../../../../app/thunk-extra-arguments';
+import {FacetSearchAPIClient} from '../../../../api/search/search-api-client';
 
 /**
  * Executes a facet search (i.e., a search for facet values in a facet search box).
@@ -24,12 +26,15 @@ import {facetIdDefinition} from '../../generic/facet-actions-validation';
 export const executeFacetSearch = createAsyncThunk<
   {facetId: string; response: FacetSearchResponse},
   string,
-  AsyncThunkSearchOptions<StateNeededForFacetSearch>
+  AsyncThunkOptions<
+    StateNeededForFacetSearch,
+    ClientThunkExtraArguments<FacetSearchAPIClient>
+  >
 >(
   'facetSearch/executeSearch',
   async (
     facetId: string,
-    {dispatch, getState, extra: {searchAPIClient, validatePayload}}
+    {dispatch, getState, extra: {apiClient, validatePayload}}
   ) => {
     const state = getState();
     let req: SpecificFacetSearchRequest | CategoryFacetSearchRequest;
@@ -43,7 +48,7 @@ export const executeFacetSearch = createAsyncThunk<
       );
     }
 
-    const response = await searchAPIClient.facetSearch(req);
+    const response = await apiClient.facetSearch(req);
     dispatch(logFacetSearch(facetId));
 
     return {facetId, response};